MAJ : Flag -Weverything

+++ src/string/ft_strchr.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strchr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gsimeon <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: gsimeon <gsimeon@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/27 19:59:00 by gsimeon           #+#    #+#             */
-/*   Updated: 2016/08/21 19:27:14 by gsimeon          ###   ########.fr       */
+/*   Updated: 2016/10/17 10:32:50 by gsimeon          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ char	*ft_strchr(const char *s, int c)
 	while (*s && *s != (char)c)
 		s++;
 	if (*s == (char)c)
-		return ((char *)s);
+		return ((char *)(unsigned long)s);
 	else
 		return (NULL);
 }
