+++ src/string/ft_strstr.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strstr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gsimeon <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: gsimeon <gsimeon@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/27 20:33:14 by gsimeon           #+#    #+#             */
-/*   Updated: 2015/11/30 16:15:40 by gsimeon          ###   ########.fr       */
+/*   Updated: 2016/10/17 10:36:05 by gsimeon          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,11 +24,11 @@ char	*ft_strstr(const char *s1, const char *s2)
 			while (s1[i] == s2[i])
 			{
 				if (!s2[++i])
-					return ((char *)s1);
+					return ((char *)(unsigned long)s1);
 			}
 			s1++;
 		}
 		return (NULL);
 	}
-	return ((char *)s1);
+	return ((char *)(unsigned long)s1);
 }
